build: fix Linux engine compile/link/runtime errors and get ctest green
- Add missing <cmath>/<algorithm> includes that MSVC pulled in transitively
- Undef GCC's predefined bare `linux` macro so the platform enumerator stays valid
- Enable CMAKE_POSITION_INDEPENDENT_CODE so static libs link into shared ones on ELF
- Feed libclang the host GCC system include dirs (conan libclang ships no builtin
  headers) and fix the dead `#elif DPLATFORM_LINUX` typo in MirrorTool
- Qualify same-named return types (Runtime::EventsObserverDyn/Observer/PlayStatus)
  to avoid GCC -Wchanges-meaning, and make Scene.h explicit specializations inline
  instead of the non-standard `static`
- Cast to qint64 in Qt JSON serializer to dodge the int64_t ambiguity on LP64
- Name the current thread via pthread_self() in NamedThread to fix the data race
  that segfaulted Runtime.Test

+++ Editor/Include/Editor/Qt/JsonSerialization.h
@@ -175,7 +175,7 @@ namespace Editor {
     struct QtJsonSerializer<int64_t> {
         static void QtJsonSerialize(QJsonValue& outJsonValue, int64_t inValue)
         {
-            outJsonValue = inValue;
+            outJsonValue = static_cast<qint64>(inValue);
         }
 
         static void QtJsonDeserialize(const QJsonValue& inJsonValue, int64_t& outValue)
@@ -191,7 +191,7 @@ namespace Editor {
     struct QtJsonSerializer<uint64_t> {
         static void QtJsonSerialize(QJsonValue& outJsonValue, uint64_t inValue)
         {
-            outJsonValue = static_cast<int64_t>(inValue);
+            outJsonValue = static_cast<qint64>(inValue);
         }
 
         static void QtJsonDeserialize(const QJsonValue& inJsonValue, uint64_t& outValue)
